RavenDB-14274 - use the CancellationToken

+++ src/Raven.Server/Documents/PeriodicBackup/BackupTask.cs
@@ -646,7 +646,8 @@ namespace Raven.Server.Documents.PeriodicBackup
                                 $"took: {totalSw.ElapsedMilliseconds:#,#;;0}ms");
                     }
 
-                    IOExtensions.RenameFile(tempBackupFilePath, backupFilePath);
+                    if (_directUploadDestination == DirectUploadDestination.Disabled)
+                        IOExtensions.RenameFile(tempBackupFilePath, backupFilePath);
 
                     status.LocalBackup.Exception = null;
                 }
@@ -654,8 +655,11 @@ namespace Raven.Server.Documents.PeriodicBackup
                 {
                     status.LocalBackup.Exception = e.ToString();
 
-                    // deleting the temp backup file if the backup failed
-                    DeleteFile(tempBackupFilePath);
+                    if (_directUploadDestination == DirectUploadDestination.Disabled)
+                    {
+                        // deleting the temp backup file if the backup failed
+                        DeleteFile(tempBackupFilePath);
+                    }
                     throw;
                 }
             }
@@ -692,7 +696,8 @@ namespace Raven.Server.Documents.PeriodicBackup
                             FolderName = folderName,
                             FileName = fileName,
                             RetentionPolicyParameters = _retentionPolicyParameters,
-                            OnProgress = AddInfo
+                            OnProgress = AddInfo,
+                            CancellationToken = TaskCancelToken.Token
                         });
 
                 default:
